requirements.txt
fastapi>=0.95.0
pydantic>=1.10
uvicorn>=0.21.1
dnspython>=2.4.2
